feat: env, uuid, exlude and fix for users routes

+++ src/users/users.controller.ts
@@ -6,6 +6,13 @@ import {
   Patch,
   Param,
   Delete,
+  UseInterceptors,
+  ClassSerializerInterceptor,
+  ParseUUIDPipe,
+  NotFoundException,
+  Put,
+  ForbiddenException,
+  HttpCode,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -16,44 +23,53 @@ import { User } from './entities/user.entity';
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
+  @UseInterceptors(ClassSerializerInterceptor)
   @Post()
   create(@Body() createUserDto: CreateUserDto) {
-    return this.usersService.create(createUserDto);
+    const { login, password } = createUserDto ?? {};
+    return this.usersService.create(new CreateUserDto(login, password));
   }
 
+  @UseInterceptors(ClassSerializerInterceptor)
   @Get()
   findAll() {
     return this.usersService.findAll();
   }
 
+  @UseInterceptors(ClassSerializerInterceptor)
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  findOne(@Param('id', new ParseUUIDPipe()) id: string) {
     const user = this.usersService.findOne(id);
     if (!user) {
-      return;
+      throw new NotFoundException();
     }
     return this.usersService.findOne(id);
   }
 
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdatePasswordDto) {
+  @UseInterceptors(ClassSerializerInterceptor)
+  @Put(':id')
+  update(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() updateUserDto: UpdatePasswordDto,
+  ) {
     const user = this.findOne(id) as User | undefined;
-
-    if (!user) {
-      return 'error';
-    }
 
     const updatedUser = this.usersService.update(user, updateUserDto);
 
     if (!updatedUser) {
-      return 'error';
+      throw new ForbiddenException();
     }
 
     return updatedUser;
   }
 
+  @HttpCode(204)
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param('id', new ParseUUIDPipe()) id: string) {
+    const isRemoved = this.usersService.remove(id);
+    if (!isRemoved) {
+      throw new NotFoundException();
+    }
     return this.usersService.remove(id);
   }
 }
